Upgraded Terrain Visuals, Traits Expanded

+++ Game/Scripts/Systems/TerrainSystem/Core/HexTile.cs
@@ -44,7 +44,7 @@ namespace Terrain {
 
             return new Vector3(
                 horiz * (this.column + this.row/2f),
-                (int) elevation_type/100,
+                (float) elevation_type/100,
                 vert * this.row
             );
         }
